backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.5.2
sqlalchemy==2.0.29
python-dotenv==1.0.1
pydantic==2.6.4
python-multipart==0.0.9
Pillow==10.3.0
httpx==0.27.0
